make interface simpler.
the "run" and "update" buttons have been removed.

if you select a config, the url is filled in (if known).  pressing
tab & return fetches the url, so is the same as update.

after fetching an url, the focus is moved to the listbox with
configs, with the newly changed config selected.  hitting return
starts that config.  this is easier than the "run" button.

at startup, the focus is placed on the listbox if there is at least
one config.  this normally means if you have started it and fetched
a config before.  if there is no config, the focus is at the url
bar.  in the listbox, the newest config (highest mtime) is selected.
so when starting, just hitting return once is usually enough.